Adds arrays as a new primitive type.
First draft of fixed-size arrays with a tentative syntax: type[size] var_name;
Array allocation is managed by the runtime (so only Interpreter for now, and badly: Only releases on program exit).
Arrays hold a type but it's not enforced yet. Each item has its own type, which is wasteful but makes the interpreter
implementation straightforward. Arrays also holds a capacity but it's currently only used by the runtime to allocate
(and assert on indexed access), it's not accessible from the language yet.
Given that the value of Variable/VariableDeclaration nodes is now fully utilized in the array case, the variable name
is stored in the associated token, which makes more sense anyway.
Contains a lot of hacks related to type propagation (& more) for the provided test to run :)

Adds an optimized fibonacci test using an array to store intermediate results.