Moved the responsability of generating calls to destructor to the Parser
This was handled by the LLVM codegen, but it did not make sense. It was also limiting our ability to reason about the destructors: Now we can mark objects as 'moved' and avoid generating a destructor call in this case. This is not a good implementation, just a first dip into actually tracking the lifetime of your local objects. I don't how how this will actually end up.